pyproject.toml
[tool.poetry]
name = "python-stack"
version = "0.1.0"
description = ""
authors = ["miragecentury <victorien.vanroye@gmail.com>"]
readme = "README.md"
packages = [
    { include = "python_stack", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
pyyaml = "^6.0.1"
reactivex = "^4.0.4"
uvicorn = "^0.29.0"
inject = "^5.2.1"
httpx = "^0.27.0"
opentelemetry-distro = "^0"
opentelemetry-instrumentation-fastapi = "^0"
opentelemetry-instrumentation-httpx = "^0"
opentelemetry-exporter-prometheus = "^0"
opentelemetry-exporter-otlp = "^1"
opentelemetry-propagator-b3 = "^1"
structlog = "^24.1.0"


[tool.poetry.group.test.dependencies]
mypy = "^1.10.0"
types-pyyaml = "^6.0.12.20240311"
pylint = "^3.2.2"
black = "^24.4.2"

[tool.poetry.extras]


[tool.poetry.dev-dependencies]
pytest = "^8.2.0"
pytest-xdist = "^3.6.1"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
python-stack-example = "python_stack.example.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "tests/units"
filterwarnings = [
    "ignore::DeprecationWarning:reactivex",
    "ignore::DeprecationWarning:importlib",
    "ignore::DeprecationWarning:opentelemetry",
]

[tool.black]
line-length = 80

target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true

[tool.pylint]
fail-under = 9.5